Improve trait aliasing impl for restricted kernel SDK
